move Group under Customer domain
