+++ src/Adapter/Customer/Group/Repository/GroupRepository.php
@@ -25,11 +25,11 @@
  */
 declare(strict_types=1);
 
-namespace PrestaShop\PrestaShop\Adapter\Group\Repository;
+namespace PrestaShop\PrestaShop\Adapter\Customer\Group\Repository;
 
 use PrestaShop\PrestaShop\Adapter\AbstractObjectModelRepository;
-use PrestaShop\PrestaShop\Core\Domain\Group\Exception\GroupNotFoundException;
-use PrestaShop\PrestaShop\Core\Domain\Group\ValueObject\GroupId;
+use PrestaShop\PrestaShop\Core\Domain\Customer\Group\Exception\GroupNotFoundException;
+use PrestaShop\PrestaShop\Core\Domain\Customer\Group\ValueObject\GroupId;
 
 /**
  * Provides methods to access Group data storage
