+++ src/PrestaShopBundle/Resources/config/services/adapter/group.yml
@@ -8,5 +8,5 @@ services:
       - '@prestashop.adapter.legacy.configuration'
       - '@=service("prestashop.adapter.legacy.context").getContext().language.id'
 
-  prestashop.adapter.group.repository.group_repository:
-      class: 'PrestaShop\PrestaShop\Adapter\Group\Repository\GroupRepository'
+  prestashop.adapter.customer.group.repository.group_repository:
+      class: 'PrestaShop\PrestaShop\Adapter\Customer\Group\Repository\GroupRepository'
