+++ src/PrestaShopBundle/Resources/config/services/adapter/product/_specific_prices.yml
@@ -62,7 +62,7 @@ services:
             - '@prestashop.adapter.product.combination.repository.combination_repository'
             - '@prestashop.adapter.currency.repository.currency_repository'
             - '@prestashop.adapter.country.repository.country_repository'
-            - '@prestashop.adapter.group.repository.group_repository'
+            - '@prestashop.adapter.customer.group.repository.group_repository'
             - '@prestashop.adapter.customer.repository.customer_repository'
             - '@prestashop.adapter.product.repository.product_repository'
